refactor(PathFinder.java): PathFinder 인터페이스를 통해 외부 라이브러리 구현체에 대한 의존성을 낮춤
