+++ src/test/java/nextstep/subway/path/domain/PathFinderTest.java
@@ -60,7 +60,7 @@ class PathFinderTest {
     @DisplayName("출발역과 도착역 사이의 최단 경로를 조회한다.")
     @Test
     void findShortestPath() {
-        PathFinder pathFinder = PathFinder.from(Arrays.asList(신분당선, 분당선, 삼호선, 일호선));
+        PathFinder pathFinder = DijkstraShortestPathFinder.from(Arrays.asList(신분당선, 분당선, 삼호선, 일호선));
 
         Path path = pathFinder.findShortestPath(양재역, 서현역);
 
@@ -73,7 +73,7 @@ class PathFinderTest {
     @DisplayName("출발역과 도착역이 같은 경우 예외가 발생한다.")
     @Test
     void findShortestPathException1() {
-        PathFinder pathFinder = PathFinder.from(Arrays.asList(신분당선, 분당선, 삼호선, 일호선));
+        PathFinder pathFinder = DijkstraShortestPathFinder.from(Arrays.asList(신분당선, 분당선, 삼호선, 일호선));
 
         Assertions.assertThatThrownBy(() -> pathFinder.findShortestPath(양재역, 양재역))
                 .isInstanceOf(PathNotFoundException.class)
@@ -83,7 +83,7 @@ class PathFinderTest {
     @DisplayName("출발역과 도착역이 연결되어 있지 않은 경우 예외가 발생한다.")
     @Test
     void findShortestPathException2() {
-        PathFinder pathFinder = PathFinder.from(Arrays.asList(신분당선, 분당선, 삼호선, 일호선));
+        PathFinder pathFinder = DijkstraShortestPathFinder.from(Arrays.asList(신분당선, 분당선, 삼호선, 일호선));
 
         Assertions.assertThatThrownBy(() -> pathFinder.findShortestPath(양재역, 소요산역))
                 .isInstanceOf(PathNotFoundException.class)
